% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptiRes.R
\name{Plot_ColorfulClusterTree}
\alias{Plot_ColorfulClusterTree}
\title{Plot Colorful Cluster Tree}
\usage{
Plot_ColorfulClusterTree(seu_obj, results_df, dims = 1:5)
}
\arguments{
\item{seu_obj}{A Seurat object containing preprocessed single-cell data.}

\item{results_df}{A data frame containing the silhouette scores and resolutions, typically the output of \code{scSilhouette}.}
}
\value{
A dendrogram object with branches and labels colored by cluster.
}
\description{
This function generates a colorful dendrogram for the Seurat object's cluster tree.
}
\examples{
\dontrun{
# Assuming 'seu_obj' is a preprocessed Seurat object and 'results_df' is the output of scSilhouette
results_df <- scSilhouette(seu_obj)
Plot_ColorfulClusterTree(seu_obj, results_df)
}
}
