% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptiRes.R
\name{scSilhouette}
\alias{scSilhouette}
\title{Calculate Average Silhouette Scores Across Resolutions}
\usage{
scSilhouette(seu_obj, resolutions = seq(0.01, 2, 0.01), verbose = TRUE)
}
\arguments{
\item{seu_obj}{A Seurat object containing preprocessed single-cell data, including UMAP embeddings.}

\item{resolutions}{Numeric vector of resolution values to evaluate. Must be positive. Default is \code{seq(0.01, 2, 0.01)}.}

\item{verbose}{Logical. Whether to print progress and results. Default is \code{TRUE}.}
}
\value{
A data frame with two columns: \code{Resolution} and \code{SilhouetteScore}, representing the resolution values and their corresponding average silhouette scores.
}
\description{
This function computes the average silhouette scores for clustering results across a range of resolutions using a Seurat object. It aids in determining the optimal clustering resolution for single-cell RNA-seq data analysis.
}
\examples{
\dontrun{
# Assuming 'seu_obj' is a preprocessed Seurat object with UMAP embeddings
results_df <- scSilhouette(seu_obj, resolutions = seq(0.1, 1.0, 0.1))
# Plot the silhouette scores
Plot_res(results_df)
}
}
