% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptiRes.R
\name{Plot_res}
\alias{Plot_res}
\title{Plot Silhouette Scores Across Resolutions}
\usage{
Plot_res(results_df)
}
\arguments{
\item{results_df}{A data frame with columns \code{Resolution} and \code{SilhouetteScore}, typically the output of \code{scSilhouette}.}
}
\value{
A ggplot object showing the silhouette scores and the optimal resolution.
}
\description{
This function visualizes the silhouette scores for different clustering resolutions and highlights the optimal resolution.
}
\examples{
\dontrun{
# Assuming 'results_df' is the output of scSilhouette
Plot_res(results_df)
}
}
